+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to Audit.NET and its extensions will be documented in this f
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [27.4.1] - 2025-02-08:
+- Audit.EntityFramework.Identity / Audit.EntityFramework.Identity.Core: Fixing issue with `AuditIdentityDbContext` when using the `SaveChanges(bool)` overload (#729)
+- Audit.WebApi.Core: Security upgrade Microsoft.AspNetCore.Mvc from 2.2.0 to 2.3.0 (#719)
+
 ## [27.4.0] - 2025-01-28:
 - Audit.NET: Changing the default implementation of `AuditDataProvider.CloneValue` to use the `Clone` method
 when the target value implements `ICloneable`